+++ examples/with_inline_policy/main.tf
@@ -2,7 +2,7 @@ module "inline_policy" {
   source             = "./../../"
   name               = "example-inline-policy-role"
   assume_role_policy = data.aws_iam_policy_document.ec2_assume_role_policy.json
-  description        = "EC2 role with variety of permissions"
+  description        = "EC2 role with a variety of permissions"
   inline_policy = [{
     name = "example_inline_policy"
     policy = jsonencode({
@@ -17,7 +17,7 @@ module "inline_policy" {
     })
     },
     {
-      name   = "boldlink-policy-10525"
+      name   = "example_inline_policy2"
       policy = data.aws_iam_policy_document.inline_policy.json
   }]
 }